% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planned_value.R
\name{get_planned_value}
\alias{get_planned_value}
\title{Calculate the Planned Value of a Project Schedule to Date}
\usage{
get_planned_value(df, project_value)
}
\arguments{
\item{df}{A Data Frame of the Project}

\item{project_value}{The total value of the project}
}
\description{
Given a set of project activities start times, end times, progress and costs,
this function calculates the Planned Value across the Project Schedule
}
\keyword{internal}
