% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\docType{data}
\name{project}
\alias{project}
\title{Dummy Project Schedule}
\format{
A data frame with 12 rows and 4 variables
\itemize{
\item \strong{start:} Start Date of the activity.
\item \strong{end:} End Date of the activity.
\item \strong{progress:} Proportion between 0 and 1 representing percentage
completed for each activity (1 = 100\% complete)
\item \strong{planned_cost:} The planned costs of each activity
}
}
\usage{
data(project)
}
\description{
A dataset of a dummy project. The variables are as follows:
}
\keyword{datasets}
